+++ resource/property.go
