+++ resource/template.go
